requirements.txt
asgiref==3.2.10
attrs==19.3.0
certifi==2020.4.5.2
chardet==3.0.4
Django==2.2.13
djangorestframework==3.11.0
djangorestframework-csv==2.1.0
drf-renderer-xlsx==0.3.7
et-xmlfile==1.0.1
idna==2.9
importlib-metadata==1.6.1
jdcal==1.4.1
more-itertools==8.4.0
openpyxl==3.0.3
packaging==20.4
pathlib2==2.3.5
pluggy==0.13.1
py==1.8.2
pyparsing==2.4.7
pytest==5.4.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
unicodecsv==0.14.1
urllib3==1.25.9
wcwidth==0.2.4
zipp==3.1.0
